Reused function createKeyRing
